Add explicit cast to size_t for guard clause > `m_DataLen`.

+++ Packet++/src/TextBasedProtocol.cpp
@@ -387,7 +387,7 @@ namespace pcpp
 
 	void TextBasedProtocolMessage::validateFieldsOffset(int offset) const
 	{
-		if (offset < 0 || offset > m_DataLen)
+		if (offset < 0 || static_cast<size_t>(offset) > m_DataLen)
 			throw std::invalid_argument(
 			    "Field offset must be greater or equal to zero and less than or equal to the data length");
 	}
